Move classes to matching files
