+++ Vulcan.NET/Key.cs
@@ -1,0 +1,137 @@
+﻿namespace Vulcan.NET
+{
+    /// <summary>
+    /// Represents all available keys on the keyboard
+    /// </summary>
+    public enum Key
+    {
+#pragma warning disable CS1591//No comments for key definitions
+        //Column 1
+        ESC = 0,
+        TILDE = 1,
+        TAB = 2,
+        CAPS_LOCK = 3,
+        LEFT_SHIFT = 4,
+        LEFT_CONTROL = 5,
+        //Column 2
+        ONE = 6,
+        Q = 7,
+        A = 8,
+        ISO_BACKSLASH = 9,//ISO only
+        LEFT_WINDOWS = 10,
+        //Column 3
+        F1 = 11,
+        TWO = 12,
+        W = 13,
+        S = 14,
+        Z = 15,
+        LEFT_ALT = 16,
+        //Column 4
+        F2 = 17,
+        THREE = 18,
+        E = 19,
+        D = 20,
+        X = 21,
+        //Column 5
+        F3 = 23,
+        FOUR = 24,
+        R = 25,
+        F = 26,
+        C = 27,
+        //Column 6
+        F4 = 28,
+        FIVE = 29,
+        T = 30,
+        G = 31,
+        V = 32,
+        //Column 7
+        SIX = 33,
+        Y = 34,
+        H = 35,
+        B = 36,
+        SPACE = 37,
+        //Column 8
+        F5 = 48,
+        SEVEN = 49,
+        U = 50,
+        J = 51,
+        N = 52,
+        //Column 9
+        F6 = 53,
+        EIGHT = 54,
+        I = 55,
+        K = 56,
+        M = 57,
+        //Column 10
+        F7 = 59,
+        NINE = 60,
+        O = 61,
+        L = 62,
+        COMMA = 63,
+        //Column 11
+        F8 = 65,
+        ZERO = 66,
+        P = 67,
+        SEMICOLON = 68,
+        PERIOD = 69,
+        RIGHT_ALT = 70,
+        //Column 12
+        MINUS = 72,
+        OPEN_BRACKET = 73,
+        APOSTROPHE = 74,
+        FORWARD_SLASH = 75,
+        FN_Key = 76,
+        //Column 13
+        F9 = 78,
+        EQUALS = 79,
+        CLOSE_BRACKET = 80,
+        BACKSLASH = 81,//ANSI only
+        RIGHT_SHIFT = 82,
+        APPLICATION_SELECT = 83,
+        //Column 14
+        F10 = 84,
+        F11 = 85,
+        F12 = 86,
+        BACKSPACE = 87,
+        ENTER = 88,
+        RIGHT_CONTROL = 89,
+        ISO_HASH = 96,//ISO ONLY
+        //Column 15
+        PRINT_SCREEN = 99,
+        INSERT = 100,
+        DELETE = 101,
+        ARROW_LEFT = 102,
+        //Column 16
+        SCROLL_LOCK = 103,
+        HOME = 104,
+        END = 105,
+        ARROW_UP = 106,
+        ARROW_DOWN = 107,
+        //Column 17
+        PAUSE_BREAK = 108,
+        PAGE_UP = 109,
+        PAGE_DOWN = 110,
+        ARROW_RIGHT = 111,
+        //Column 18
+        NUM_LOCK = 113,
+        NUM_SEVEN = 114,
+        NUM_FOUR = 115,
+        NUM_ONE = 116,
+        NUM_ZERO = 117,
+        //Column 19
+        NUM_SLASH = 119,
+        NUM_EIGHT = 120,
+        NUM_FIVE = 121,
+        NUM_TWO = 122,
+        //Column 20
+        NUM_ASTERISK = 124,
+        NUM_NINE = 125,
+        NUM_SIX = 126,
+        NUM_THREE = 127,
+        NUM_PERIOD = 128,
+        //Column 21
+        NUM_MINUS = 129,
+        NUM_PLUS = 130,
+        NUM_ENTER = 131
+    }
+}
